requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
numpy>=1.21.0
pytest>=6.2.5
pydantic>=1.8.2
sentence-transformers>=2.2.0
pandas>=1.3.0
torch>=1.3.0
faiss-cpu>=1.7.0
tqdm>=4.65.0
transformers>=4.30.0
datasets>=2.12.0
openai>=1.0.0
python-dotenv>=0.19.0
tenacity>=8.0.1
sentencepiece>=0.1.99